Refactor warmup scripts. Executable scripts now in bin/ and common code in warmup package. Remove outlier summary script. Fixes #150.
