+++ warmup/krun_results.py
@@ -1,0 +1,19 @@
+import bz2
+import json
+
+
+def read_krun_results_file(results_file):
+    """Return the JSON data stored in a Krun results file.
+    """
+    results = None
+    with bz2.BZ2File(results_file, 'rb') as file_:
+        results = json.loads(file_.read())
+        return results
+    return None
+
+
+def write_krun_results_file(results, filename):
+    """Write a Krun results file to disk.
+    """
+    with bz2.BZ2File(filename, 'wb') as file_:
+        file_.write(json.dumps(results))
