+++ warmup/plotting.py
@@ -1,0 +1,49 @@
+SPINE_LINESTYLE = "solid"
+SPINE_LINEWIDTH = 1
+
+
+def add_margin_to_axes(axis, x=0.01, y=0.01):
+    """Seaborn-friendly way to add margins to axes (default 1% margin).
+    """
+
+    if x > .0:
+        xlim = axis.get_xlim()
+        xmargin = (xlim[1] - xlim[0]) * x
+        axis.set_xlim(xlim[0] - xmargin, xlim[1] + xmargin)
+    if y > .0:
+        ylim = axis.get_ylim()
+        ymargin = (ylim[1] - ylim[0]) * y
+        axis.set_ylim(ylim[0] - ymargin, ylim[1] + ymargin)
+
+
+def compute_grid_offsets(d_min, d_max, num):
+    rng = float(d_max) - d_min
+    freq =  rng / num
+    return [d_min + i * freq for i in xrange(num + 1)]
+
+
+def style_axis(ax, major_xticks, minor_xticks, major_yticks, minor_yticks, tick_fontsize):
+    ax.set_xticks(major_xticks)
+    ax.set_xticks(minor_xticks, minor=True)
+    ax.set_yticks(major_yticks)
+    ax.set_yticks(minor_yticks, minor=True)
+
+    x_ax = ax.get_xaxis()
+    y_ax = ax.get_yaxis()
+
+    x_ax.set_ticks_position('none')
+    y_ax.set_ticks_position('none')
+    x_ax.set_tick_params(labelsize=tick_fontsize)
+    y_ax.set_tick_params(labelsize=tick_fontsize)
+
+    ax.grid(which='minor', alpha=0.4)
+    ax.grid(which='major', alpha=0.8)
+
+    for i in ["top", "bottom"]:
+        ax.spines[i].set_linestyle(SPINE_LINESTYLE)
+        ax.spines[i].set_linewidth(SPINE_LINEWIDTH)
+
+    for i in ["right", "left"]:
+        ax.spines[i].set_visible(False)
+
+    ax.frameon = False
